projects/pdf_qa_bot/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
flask>=3.0.0
pdfminer.six>=20221105

